backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
google-genai>=1.0.0
pypdf>=4.0.0
google-auth>=2.0.0
python-decouple==3.8
Pillow==10.1.0
gunicorn==21.2.0
social-auth-app-django==5.4.0
django-filter==23.3
whitenoise
